+++ dev/obsolete/PyOctaveBand.py
