+++ dev/obsolete/pipeline_cnn_ext_obsolete.py
