+++ dev/obsolete/pipeline_cnn_obsolete.py
